Split Reservations Domain into slices
