--- Sample/Tickets/Tickets.Tests/Reservations/ConfirmReservationTests.cs
@@ -1,35 +1,0 @@
-using FluentAssertions;
-using Tickets.Reservations;
-using Tickets.Reservations.Events;
-using Tickets.Tests.Builders;
-using Tickets.Tests.Extensions;
-using Xunit;
-
-namespace Tickets.Tests.Reservations
-{
-    public class ConfirmReservationTests
-    {
-        [Fact]
-        public void ForTentativeReservation_ShouldSucceed()
-        {
-            // Given
-            var reservation = ReservationBuilder
-                .Create()
-                .Tentative()
-                .Build();
-
-            // When
-            reservation.Confirm();
-
-            // Then
-            reservation.Status.Should().Be(ReservationStatus.Confirmed);
-            reservation.Version.Should().Be(2);
-
-            var @event = reservation.PublishedEvent<ReservationConfirmed>();
-
-            @event.Should().NotBeNull();
-            @event.Should().BeOfType<ReservationConfirmed>();
-            @event!.ReservationId.Should().Be(reservation.Id);
-        }
-    }
-}
