+++ Sample/Tickets/Tickets.Tests/Stubs/Reservations/FakeReservationNumberGenerator.cs
@@ -1,5 +1,6 @@
 using System;
 using Tickets.Reservations;
+using Tickets.Reservations.NumberGeneration;
 
 namespace Tickets.Tests.Stubs.Reservations
 {
