--- Sample/Tickets/Tickets/Concerts/ConcertCommandHandler.cs
@@ -1,6 +1,0 @@
-namespace Tickets.Concerts
-{
-    public class ConcertCommandHandler
-    {
-    }
-}
