--- Sample/Tickets/Tickets/Concerts/ConcertQueryHandler.cs
@@ -1,7 +1,0 @@
-namespace Tickets.Concerts
-{
-    public class ConcertQueryHandler
-    {
-        
-    }
-}
