+++ Sample/Tickets/Tickets/Reservations/ConfirmingReservation/ReservationConfirmed.cs
@@ -1,0 +1,22 @@
+using System;
+using Core.Events;
+using Newtonsoft.Json;
+
+namespace Tickets.Reservations.ConfirmingReservation
+{
+    public class ReservationConfirmed : IEvent
+    {
+        public Guid ReservationId { get; }
+
+        [JsonConstructor]
+        private ReservationConfirmed(Guid reservationId)
+        {
+            ReservationId = reservationId;
+        }
+
+        public static ReservationConfirmed Create(Guid reservationId)
+        {
+            return new(reservationId);
+        }
+    }
+}
