--- Sample/Tickets/Tickets/Reservations/Events/ReservationCancelled.cs
@@ -1,22 +1,0 @@
-using System;
-using Core.Events;
-using Newtonsoft.Json;
-
-namespace Tickets.Reservations.Events
-{
-    public class ReservationCancelled : IEvent
-    {
-        public Guid ReservationId { get; }
-
-        [JsonConstructor]
-        private ReservationCancelled(Guid reservationId)
-        {
-            ReservationId = reservationId;
-        }
-
-        public static ReservationCancelled Create(Guid reservationId)
-        {
-            return new ReservationCancelled(reservationId);
-        }
-    }
-}
