+++ Sample/Tickets/Tickets/Reservations/NumberGeneration/ReservationNumberGenerator.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace Tickets.Reservations.NumberGeneration
+{
+    public interface IReservationNumberGenerator
+    {
+        string Next();
+    }
+
+    internal class ReservationNumberGenerator: IReservationNumberGenerator
+    {
+        public string Next() => Guid.NewGuid().ToString();
+    }
+}
