--- Sample/Tickets/Tickets/Reservations/Queries/GetReservationById.cs
@@ -1,10 +1,0 @@
-using System;
-using Core.Queries;
-using Tickets.Reservations.Projections;
-
-namespace Tickets.Reservations.Queries
-{
-    public record GetReservationById(
-        Guid ReservationId
-    ): IQuery<ReservationDetails>;
-}
