--- Sample/Tickets/Tickets/Reservations/ReservationNumberGenerator.cs
@@ -1,14 +1,0 @@
-using System;
-
-namespace Tickets.Reservations
-{
-    public interface IReservationNumberGenerator
-    {
-        string Next();
-    }
-
-    internal class ReservationNumberGenerator: IReservationNumberGenerator
-    {
-        public string Next() => Guid.NewGuid().ToString();
-    }
-}
