--- Sample/Tickets/Tickets/Seats/SeatCommandHandler.cs
@@ -1,7 +1,0 @@
-namespace Tickets.Seats
-{
-    public class SeatCommandHandler
-    {
-        
-    }
-}
