--- Sample/Tickets/Tickets/Seats/SeatQueryHandler.cs
@@ -1,7 +1,0 @@
-namespace Tickets.Seats
-{
-    public class SeatsQueryHandler
-    {
-        
-    }
-}
