--- Sample/Tickets/Tickets/Tickets/TicketCommandHandler.cs
@@ -1,7 +1,0 @@
-namespace Tickets.Tickets
-{
-    public class TicketCommandHandler
-    {
-        
-    }
-}
