--- Sample/Tickets/Tickets/Tickets/TicketQueryHandler.cs
@@ -1,7 +1,0 @@
-namespace Tickets.Tickets
-{
-    public class TicketQueryHandler
-    {
-        
-    }
-}
